gitObjectType
